Revert "Fix compile cast warnings present in VS"
This reverts commit 899af3a7d6ad2535314b870f56dcf12a6b47126f.

+++ engine/source/platformWin32/winDirectInput.cc
@@ -732,7 +732,7 @@ inline void DInputManager::fireXInputButtonEvent( int controllerID, bool forceFi
 	  Con::printf("%s", objName);
 	  */
 
-      buildXInputEvent( controllerID, XI_BUTTON, objInst, action, (float) ( action == XI_MAKE ? 1 : 0 ) );
+      buildXInputEvent( controllerID, XI_BUTTON, objInst, action, ( action == XI_MAKE ? 1 : 0 ) );
    }
 }
